something wrong on rendering(calculating line len)

+++ includes/structs/ray.h
@@ -6,7 +6,7 @@
 /*   By: konsolka <konsolka@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/27 21:38:59 by user              #+#    #+#             */
-/*   Updated: 2020/04/28 11:37:05 by konsolka         ###   ########.fr       */
+/*   Updated: 2020/04/28 18:47:53 by konsolka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ typedef struct	s_ray
 {
 	float		x;
 	float		y;
-	float		len;
+	double		len;
 	float		pov;
 	int			wall_c;
 	int			angle;
